new project layout and demo rework
