+++ demo/Dockerfile
@@ -1,15 +1,6 @@
 FROM python:2-alpine
 RUN mkdir /demo
-COPY sendmessage.py /demo/
+COPY . /demo/
 RUN pip install pybabblesdk
-ENTRYPOINT ["python", "/demo/sendmessage.py"]
+ENTRYPOINT ["python", "/demo/run.py"]
 CMD []
-
-#FROM python:2-alpine
-#RUN mkdir /demo && mkdir /demo/pybabblesdk
-#COPY pybabblesdk /demo/pybabblesdk
-#COPY sendmessage.py /demo/
-#ENV PYTHONPATH=$PYTHONPATH:/demo/pybabblesdk
-#RUN python -m pip install six
-#ENTRYPOINT ["python", "/demo/sendmessage.py"]
-#CMD []
